+++ spec/unit/generator/prefix_header_spec.rb
@@ -21,10 +21,10 @@ module Pod
       EOS
     end
 
+    # @note Declaring a subspec was found in issue #1449 to generate duplicates of the prefix_header_contents
     it "does not duplicate the contents of the specification's prefix header when a subspec is declared" do
       @spec.prefix_header_contents = '#import "BlocksKit.h"'
       @spec.prefix_header_file = nil
-      # Declaring a subspec was found in issue #1449 to generate duplicates of the prefix_header_contents
       @spec.subspec 'UI' do |subspec|
         subspec.source_files = 'Source/UI/*.{h,m}'
       end
@@ -37,10 +37,10 @@ module Pod
       EOS
     end
 
+    # @note Declaring a subspec was found in issue #1449 to generate duplicates of the prefix_header_contents
     it "does not duplicate the contents of the specification's prefix header when a subspec is declared multiple times" do
       @spec.prefix_header_contents = '#import "BlocksKit.h"'
       @spec.prefix_header_file = nil
-      # Declaring a subspec was found in issue #1449 to generate duplicates of the prefix_header_contents
       @spec.subspec 'UI' do |su|
           su.source_files = 'Source/UI/*.{h,m}'
       end
@@ -48,15 +48,7 @@ module Pod
       @spec.subspec 'Helpers' do |sh|
           sh.source_files = 'Source/Helpers/*.{h,m}'
       end
-    
-      @spec.subspec 'Additions' do |sa|
-          sa.source_files = 'Source/Additions/*.{h,m}'
-      end
-    
-      @spec.subspec 'Dashboard' do |sd|
-          sd.source_files = 'Source/Dashboard/*.{h,m}'
-          sd.resources    = 'Source/Dashboard/*.{xib}'
-      end
+
       @gen.generate.should == <<-EOS.strip_heredoc
       #ifdef __OBJC__
       #import <UIKit/UIKit.h>